Refact: 프로젝트 디테일 로직 분리 적용
